Just for being updated

+++ app/src/main/java/code/name/monkey/retromusic/volume/AudioVolumeContentObserver.java
@@ -23,13 +23,16 @@ import androidx.annotation.NonNull;
 public class AudioVolumeContentObserver extends ContentObserver {
 
     private final OnAudioVolumeChangedListener mListener;
+
     private final AudioManager mAudioManager;
+
     private final int mAudioStreamType;
-    private int mLastVolume;
+
+    private float mLastVolume;
 
     AudioVolumeContentObserver(@NonNull Handler handler, @NonNull AudioManager audioManager,
-                               int audioStreamType,
-                               @NonNull OnAudioVolumeChangedListener listener) {
+            int audioStreamType,
+            @NonNull OnAudioVolumeChangedListener listener) {
 
         super(handler);
         mAudioManager = audioManager;
@@ -44,8 +47,8 @@ public class AudioVolumeContentObserver extends ContentObserver {
     @Override
     public void onChange(boolean selfChange, Uri uri) {
         if (mAudioManager != null && mListener != null) {
-            int maxVolume = mAudioManager.getStreamMaxVolume(mAudioStreamType);
-            int currentVolume = mAudioManager.getStreamVolume(mAudioStreamType);
+            float maxVolume = mAudioManager.getStreamMaxVolume(mAudioStreamType);
+            float currentVolume = mAudioManager.getStreamVolume(mAudioStreamType);
             if (currentVolume != mLastVolume) {
                 mLastVolume = currentVolume;
                 mListener.onAudioVolumeChanged(currentVolume, maxVolume);
